fix(precision): 白银 2dp 分制 + 伤害小数化（全程不取整）
- Silver.cs（新，Link 进 tests）：白银分制精度——2 位小数金额，收支不再取整损失。
- Merchant（Trade/BuyList/Panel/Shelf + Sell/Trade 测试）：价格/收支走 Silver 2dp 分制。
- 伤害小数化：CombatData/Duel/CombatResolver/CombatResult/Effects 伤害值 int→double 全程
  不取整（期望微降 0.3~0.5，Sim 基线重核见 body-parts commit）；ActorSprite/CombatLogFormatter/
  CombatMoteText 展示按小数呈现。CombatResolverTests 断言同步。
- csproj：Silver + InfectionCourseLogic Link。

数值精度通则见设计文档 §2。伤害微降幅度拟定待调（Sim 校准）。

+++ godot/scripts/ActorSprite.cs
@@ -140,7 +140,7 @@ public sealed partial class ActorSprite : Node2D
         }
 
         // 伤害强度归一（拟定待调）：常规伤害约 0~30 铺满 0..1，断肢直接拉满。
-        float sev = Mathf.Clamp(hit.Damage / 30f, 0f, 1f);
+        float sev = Mathf.Clamp((float)(hit.Damage / 30.0), 0f, 1f);
         if (hit.Severed)
         {
             sev = 1f;
